# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/category/": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "Gets the list of all available categories. (Use token from Login response) <br>Authorization format example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZW1haWwiOiJ1c2VyQGV4YW1wbGUuY29tIiwiaWF0IjoxNjk3ODEwMzAyLCJleHAiOjE2O",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of categories"
          },
          "401": {
            "description": "Unauthorized: Invalid token"
          },
          "404": {
            "description": "No categories found"
          },
          "500": {
            "description": "An error occurred while fetching categories"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "description": "Creates a new category.(Use token from Login response) <br>Authorization format example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZW1haWwiOiJ1c2VyQGV4YW1wbGUuY29tIiwiaWF0IjoxNjk3ODEwMzAyLCJleHAiOjE2O",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/PostCategory"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created category"
          },
          "400": {
            "description": "Missing required property"
          },
          "401": {
            "description": "Unauthorized: Invalid token"
          },
          "500": {
            "description": "An error occurred while creating category"
          }
        }
      }
    },
    "/category/{id}": {
      "put": {
        "tags": [
          "Categories"
        ],
        "description": "Updates a category based on their id.(Use token from Login response) <br>Authorization format example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZW1haWwiOiJ1c2VyQGV4YW1wbGUuY29tIiwiaWF0IjoxNjk3ODEwMzAyLCJleHAiOjE2O",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/PutCategory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category updated successfully"
          },
          "401": {
            "description": "Unauthorized: Invalid token"
          },
          "404": {
            "description": "Category not found"
          },
          "500": {
            "description": "An error occurred while updating category"
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "description": "Deletes a category based on their id. (Use token from Login response) <br>Authorization format example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwiZW1haWwiOiJ1c2VyQGV4YW1wbGUuY29tIiwiaWF0IjoxNjk3ODEwMzAyLCJleHAiOjE2O",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Category updated successfully"
          },
          "401": {
            "description": "Unauthorized: Invalid token"
          },
          "404": {
            "description": "Category not found"
          },
          "500": {
            "description": "An error occurred while deleting category"
          }
        }
      }
    },
...
  },
  "definitions": {
    "PostCategory": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Work"
        }
      },
      "required": [
        "name"
      ]
    },
...
    "PutCategory": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Home"
        }
      },
      "required": [
        "name"
      ]
    },
...
  }
}